Core: add support to gather stats from child processes.
Some web servers like NginX spawn child processes per configuration,
as they are not based on threads its required to track the resources
usage for all of them when running a benchmark in a different way.

Starting from this patch, when giving the primary PID to monitor, it
will lookup for all childs and add them to the watch list.

Signed-off-by: Eduardo Silva <edsiper@gmail.com>